requirements.txt
pygame-ce
moderngl
flask
pygbag
buildozer
cython
javac
cmake
pillow
pyinstaller
